chore(merkle_root): move function into lib.rs
